+++ packages/flutter_parameterized_test/example/flutter_parameterized_test_example.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_parameterized_test/flutter_parameterized_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:parameterized_source/parameterized_source.dart';
 
 void main() {
   parameterizedTest(
@@ -10,7 +9,7 @@ void main() {
       'apple, 5',
       'banana, 6',
     ]),
-        (List<dynamic> values) {
+    (List<dynamic> values) {
       final String input = values[0];
       final expected = values[1];
 
@@ -27,7 +26,7 @@ void main() {
       ['apple', 5],
       ['banana', 6],
     ]),
-        (List<dynamic> values) {
+    (List<dynamic> values) {
       final String input = values[0];
       final expected = values[1];
 
@@ -44,7 +43,7 @@ void main() {
       'bike',
       'four',
     ]),
-        (List<dynamic> values) {
+    (List<dynamic> values) {
       final String input = values[0];
 
       final actual = input.length;