+++ packages/flutter_parameterized_test/lib/flutter_parameterized_test.dart
@@ -1,4 +1,7 @@
 library flutter_parameterized_test;
 
-export '/src/flutter_parameterized_test_base.dart';
+@Deprecated('Removed in v1.0.0')
 export 'package:parameterized_source/parameterized_source.dart';
+
+@Deprecated('Removed in v1.0.0')
+export '/src/flutter_parameterized_test_base.dart';