+++ packages/parameterized_source/lib/parameterized_source.dart
@@ -1,3 +1,4 @@
 library parameterized_source;
 
+@Deprecated('Removed in v1.0.0')
 export 'src/parameterized_source.dart';