+++ packages/parameterized_test/lib/src/parameterized_test_base.dart
@@ -1,39 +1,44 @@
 import 'package:meta/meta.dart';
-import 'package:parameterized_source/parameterized_source.dart';
 import 'package:test/test.dart';
 
-/// Create a group of parameterized tests.
+import 'test_options/group_test_options.dart';
+import 'test_parameters.dart';
+import 'value_source.dart';
+
+/// Create a new parameterizedTest with given [description], [values] and [body]
 ///
-/// The group will iterate over the supplied [parameters].
-/// The [parameters] are supplied through a [ParameterizedSource].
-/// Each test will be executed with the next parameter which will receive the parameter in the [body].
-///
-/// parameterized test also have the same options as group tests have. These options will be passed to the group function.
+/// [parameterizedTest] also have the same options as group tests have. These options will be passed to the group function.
 ///
 /// For example:
 /// ```dart
 /// parameterizedTest(
-///     'Amount of letters',
-///     ParameterizedSource.csv([
-///       'kiwi, 4',
-///       'apple, 5',
-///       'banana, 6',
-///     ]),
-///     (List<dynamic> values) {
-///       final String input = values[0];
-///       final expected = values[1];
-///
-///       final actual = input.length;
-///
-///       expect(actual, expected);
-///     },
-///   );
+///   'Amount of letters',
+///   [
+///     ['kiwi', 4],
+///     ['apple', 5],
+///     ['banana', 6].withTestOptions(skip: 'skip this'),
+///   ],
+///   p2((String word, int length) {
+///     expect(word.length, length);
+///   }),
+/// );
 /// ```
 @isTestGroup
+//ignore: long-parameter-list
 void parameterizedTest(
+  /// Test description.
   Object description,
-  ParameterizedSource parameters,
-  void Function(List<dynamic> values) body, {
+
+  /// List of test values. For each values in the list a test will be executed.
+  Iterable<dynamic> values,
+
+  /// The test body which is executed for each test value.
+  /// See [TestParameters] for more info on different bodies.
+  TestParameters body, {
+  dynamic Function()? setUp,
+
+  /// Provide a tearDown function to the `group` test.
+  dynamic Function()? tearDown,
   String? testOn,
   Timeout? timeout,
   dynamic skip,
@@ -41,18 +46,77 @@ void parameterizedTest(
   Map<String, dynamic>? onPlatform,
   int? retry,
 }) {
-  group(description, () {
-    for (final parameter in parameters.params) {
-      test(
-        parameter,
-        () => body(parameter),
-        testOn: testOn,
-        timeout: timeout,
-        skip: skip,
-        tags: tags,
-        onPlatform: onPlatform,
-        retry: retry,
-      );
-    }
-  });
+  ValueSource(
+    values,
+    GroupTestOptions(
+      description: description,
+      setUp: setUp,
+      tearDown: tearDown,
+      testOn: testOn,
+      timeout: timeout,
+      skip: skip,
+      tags: tags,
+      onPlatform: onPlatform,
+      retry: retry,
+    ),
+  ).executeTests(body);
+}
+
+/// Create a new parameterizedGroup with given [description], [values] and [body]
+///
+/// [parameterizedGroup] also have the same options as group tests have. These options will be passed to the group function.
+///
+/// For example:
+/// ```dart
+/// parameterizedGroup(
+///   'Amount of letters',
+///   [
+///     ['kiwi', 4],
+///     ['apple', 5],
+///     ['banana', 6].withTestOptions(skip: 'skip this'),
+///   ],
+///   p2((String word, int length) {
+///     test('test word length',() {
+///       expect(word.length, length);
+///       });
+///   }),
+/// );
+/// ```
+@isTestGroup
+//ignore: long-parameter-list
+void parameterizedGroup(
+  /// Group description.
+  Object description,
+
+  /// List of group values. For each values in the list a group test will be executed.
+  Iterable<dynamic> values,
+
+  /// The test body which is executed for each group value.
+  /// See [TestParameters] for more info on different bodies.
+  TestParameters body, {
+  dynamic Function()? setUp,
+
+  /// Provide a tearDown function to the `group` test.
+  dynamic Function()? tearDown,
+  String? testOn,
+  Timeout? timeout,
+  dynamic skip,
+  dynamic tags,
+  Map<String, dynamic>? onPlatform,
+  int? retry,
+}) {
+  ValueSource(
+    values,
+    GroupTestOptions(
+      description: description,
+      setUp: setUp,
+      tearDown: tearDown,
+      testOn: testOn,
+      timeout: timeout,
+      skip: skip,
+      tags: tags,
+      onPlatform: onPlatform,
+      retry: retry,
+    ),
+  ).executeGroup(body);
 }