+++ packages/parameterized_test/lib/src/test_source.dart
@@ -1,0 +1,6 @@
+import 'test_parameters.dart';
+
+abstract class TestSource {
+  void executeTests(TestParameters body);
+  void executeGroup(TestParameters body);
+}