+++ packages/parameterized_test/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.0.0-rc2
+
+- Added `parameterizedGroup`.
+
+## 1.0.0-rc1
+
+- DISCONTINUED `parameterized_source` package
+  - DEPRECATED `ParameterizedSource.csv`
+  - DEPRECATED `ParameterizedSource.value`
+  - DEPRECATED `ParameterizedSource.values`
+- DISCONTINUED `flutter_parameterized_test` package
+- Added new TestParameters class for providing the test body to execute.The package has
+  implementation for TestParameters support up to 10 arguments.
+- Added extensions `withTestOptions` for providing test options to test values.
+
 ## 0.2.1
 
 - Fixed issues with wrong dependencies